markdown_live_preview: Drop casts and a clone that clippy now rejects
`./script/clippy` passes `--deny warnings`, so the 13 `unnecessary_cast`
lints and one `redundant_clone` below were all fatal.

The casts are upstream drift, not bad code. `Node::child_count()`
returned `usize` in tree-sitter 0.26, so `as u32` was required when this
was written. Upstream's `2040e0de59` ("treesitter: Worker-pinned
treesitter parsing", #62784) moved the workspace pin to 0.27.0, where it
returns `u32`, and every cast became a no-op. Nothing conflicted and
nothing failed to compile — exactly the silent, conflict-free drift
CLAUDE.md warns about in the upstream-merge section.

`frontmatter_range` is cloned at its last use in `tests.rs`, so the
clone is dead.

Verified that the remaining `as u32` casts in this crate are genuine:
`start_position().row` is `usize` and the image drag width is a float.

+++ crates/markdown_live_preview/src/markdown_live_preview.rs
@@ -5159,7 +5159,7 @@ impl Extraction<'_> {
     /// neighbors and display-style layout cannot fit within a line.
     fn latex(&mut self, node: tree_sitter::Node) {
         let mut delimiters = Vec::new();
-        for index in 0..node.child_count() as u32 {
+        for index in 0..node.child_count() {
             let Some(child) = node.child(index) else {
                 continue;
             };
@@ -5410,7 +5410,7 @@ impl Extraction<'_> {
     fn list_item_markers(&mut self, node: tree_sitter::Node) {
         let mut list_marker = None;
         let mut task_marker = None;
-        for index in 0..node.child_count() as u32 {
+        for index in 0..node.child_count() {
             let Some(child) = node.child(index) else {
                 continue;
             };
@@ -5476,7 +5476,7 @@ impl Extraction<'_> {
                         "emphasis" => self.italic.push(range),
                         _ => self.bold.push(range),
                     }
-                    for index in 0..node.child_count() as u32 {
+                    for index in 0..node.child_count() {
                         let Some(child) = node.child(index) else {
                             continue;
                         };
@@ -5491,7 +5491,7 @@ impl Extraction<'_> {
                 }
                 "code_span" => {
                     self.code_spans.push(node.byte_range());
-                    for index in 0..node.child_count() as u32 {
+                    for index in 0..node.child_count() {
                         let Some(child) = node.child(index) else {
                             continue;
                         };
@@ -5503,7 +5503,7 @@ impl Extraction<'_> {
                 "inline_link" | "full_reference_link" | "collapsed_reference_link" => {
                     let mut open_bracket = None;
                     let mut close_bracket = None;
-                    for index in 0..node.child_count() as u32 {
+                    for index in 0..node.child_count() {
                         let Some(child) = node.child(index) else {
                             continue;
                         };
@@ -5531,13 +5531,13 @@ impl Extraction<'_> {
                     // literal text (the preview pane has the same limit). The
                     // image sits under a `link_text` node, not directly under
                     // the link.
-                    let wrapped_image = (0..node.child_count() as u32)
+                    let wrapped_image = (0..node.child_count())
                         .filter_map(|index| node.child(index))
                         .find_map(|child| {
                             if child.kind() == "image" {
                                 Some(child)
                             } else if child.kind() == "link_text" {
-                                (0..child.child_count() as u32)
+                                (0..child.child_count())
                                     .filter_map(|index| child.child(index))
                                     .find(|grandchild| grandchild.kind() == "image")
                             } else {
@@ -5583,7 +5583,7 @@ impl Extraction<'_> {
                     }
                     if self.is_alone_on_line(node) {
                         self.image_block(node);
-                    } else if let Some(description) = (0..node.child_count() as u32)
+                    } else if let Some(description) = (0..node.child_count())
                         .filter_map(|index| node.child(index))
                         .find(|child| child.kind() == "image_description")
                     {
@@ -5986,14 +5986,14 @@ impl Extraction<'_> {
     }
 
     fn image_alt(&self, image_node: tree_sitter::Node) -> Option<&str> {
-        let description = (0..image_node.child_count() as u32)
+        let description = (0..image_node.child_count())
             .filter_map(|index| image_node.child(index))
             .find(|child| child.kind() == "image_description")?;
         self.text.get(description.byte_range())
     }
 
     fn image_destination(&self, image_node: tree_sitter::Node) -> Option<String> {
-        let destination = (0..image_node.child_count() as u32)
+        let destination = (0..image_node.child_count())
             .filter_map(|index| image_node.child(index))
             .find(|child| child.kind() == "link_destination")?;
         self.text
@@ -6147,7 +6147,7 @@ fn toggle_callout(
 }
 
 fn push_children<'a>(node: tree_sitter::Node<'a>, stack: &mut Vec<tree_sitter::Node<'a>>) {
-    for index in (0..node.child_count() as u32).rev() {
+    for index in (0..node.child_count()).rev() {
         if let Some(child) = node.child(index) {
             stack.push(child);
         }
@@ -6155,7 +6155,7 @@ fn push_children<'a>(node: tree_sitter::Node<'a>, stack: &mut Vec<tree_sitter::N
 }
 
 fn heading_level(node: tree_sitter::Node) -> u32 {
-    for index in 0..node.child_count() as u32 {
+    for index in 0..node.child_count() {
         if let Some(child) = node.child(index) {
             match child.kind() {
                 "atx_h1_marker" => return 1,
